+++ src/components/mod.rs
@@ -1,3 +1,5 @@
+mod b_modal;
 mod b_navbar;
 
+pub use b_modal::*;
 pub use b_navbar::*;